Specifies mime type for file to download
There can be more than one VideoUrl with MediaType MP4. The one that makes the most sense to download is the one with MimeType of `video/mp4`, but there can also be one with `audio/mpegurl` which gives a (very small) `.m3u8` file.

This change just limits to the one MimeType